Alex added comments; XYu changed pictures.
