+++ package.json
@@ -35,8 +35,7 @@
     "webpack-merge": "^4.2.2"
   },
   "devDependencies": {
-	"easystarjs": "^0.4.4",
+    "easystarjs": "^0.4.4",
     "mocha": "^9.2.0"
   }
 }
-
