+++ src/assets/json/keys.json
@@ -1,0 +1,62 @@
+{
+	"bullet":{
+		"image":"string",
+		"size":"array",
+		"damage":"integer",
+		"damageType":"string",
+		"AOE":"float"
+	},
+	"enemy":{
+		"image":"string",
+		"size":"array",
+		"speed":"float",
+		"hp":"integer",
+		"value":"integer"
+	},	
+	"selection":{
+		"name":"string",
+		"picture":"string",
+		"description":"string"
+	},
+	"startParams":{
+		"lives":"integer",
+		"money":"integer",
+		"score":"integer",
+		"speedSetting":"integer",
+		"isPlaying":"boolean"
+	},
+	"enemyWaves":{
+		"waves":"dictionary",
+		"wavePause":"float"
+	},
+	"map":{
+		"tiles":"nestedArray",
+		"tileset":"string",
+		"size":"integer",
+		"path":"array",
+		"origin":"integer",
+		"dest":"integer",
+		"site":"integer"
+	},
+	"level":{
+		"selection":"dictionary",
+		"startParams":"dictionary",
+		"enemyWaves":"dictionary",
+		"map":"dictionary"
+	},
+	"tower":{
+		"name":"string",
+		"image":"string",
+		"bulletKey":"string",
+		"range":"float",
+		"speed":"float",
+		"cost":"integer",
+		"upgradeKey":"conditionalString"
+	},
+	"wave":{
+		"enemy":"string",
+		"rate":"float",
+		"quant":"integer"
+	}
+	
+}
