+++ src/assets/json/levels/0.json
@@ -1,0 +1,48 @@
+{
+	"3":{
+		"selection":{
+			"name": "test0",
+			"picture": "one",
+			"description": "Test Description 0"
+		},
+		"startParams":{
+			"lives":20,
+			"money":10,
+			"score":0,
+			"speedSetting":1,
+			"isPlaying":true
+		},
+		"enemyWaves":{
+			"waves":[
+				{
+					"enemy": "resources.enemies[\"enemy_1\"]",
+					 "rate": 0.5,
+					 "quant": 10
+				},
+				{
+					"enemy": "resources.enemies[\"enemy_2\"]",
+				 	"rate": 0.7,
+					"quant": 7
+				}
+			
+			],
+			"wavePause":5
+		},
+		"map":{
+			"tiles":[
+				[0,2,3,1,10,10,3,"a"],
+				[0,2,5,6,7,4,5,1],
+				[4,5,6,1,0,2,9,1],
+				[7,10,3,10,10,10,3,1],
+				[6,5,4,10,10,10,10,1],
+				[0,10,6,7,5,8,4, 5]
+			],
+			"tileset":"tileset",
+			"size":100,
+			"path":[4,5,6,7,8,9],
+			"origin":8,
+			"dest":9,
+			"site":10
+		}
+	}
+}
