+++ src/assets/json/levels/2.json
@@ -36,6 +36,7 @@
 				[6,5,4,10,10,10,10,1],
 				[0,10,6,7,5,4,8,1]
 			],
+			"tileset":"tileset",
 			"size":100,
 			"path":[4,5,6,7,8,9],
 			"origin":8,
