+++ src/assets/json/towers/3.json
@@ -3,7 +3,7 @@
 		"name":"Sniper Tower 1",
 		"image":"tower2a",
 		"bulletKey":"resources.bullets[\"bullet_2\"]",
-		"range":300,
+		"range":3,
 		"speed":1,
 		"cost":7,
 		"upgradeKey":"resources.towers[\"tower_2b\"]"
