+++ src/assets/json/towers/5.json
@@ -3,7 +3,7 @@
 		"name":"AOE Tower 1",
 		"image":"tower3a",
 		"bulletKey":"resources.bullets[\"bullet_3\"]",
-		"range":150,
+		"range":1.5,
 		"speed":1.5,
 		"cost":7,
 		"upgradeKey":"resources.towers[\"tower_3b\"]"
