+++ src/assets/json/towers/6.json
@@ -3,9 +3,8 @@
 		"name":"AOE Tower 2",
 		"image":"tower3b",
 		"bulletKey":"resources.bullets[\"bullet_3\"]",
-		"range":200,
+		"range":2,
 		"speed":2,
-		"cost":10,
-		"upgradeKey":null
+		"cost":10
 	}
 }
