+++ src/play objects/gameStats.js
@@ -1,64 +1,56 @@
-import ten from '../assets/images/10.png'
-
 class GameStats extends Phaser.Scene
 {
 	constructor ()
     {
 //the scene changes to this one when this keyword is used
         super('gameStats');
-		Phaser.Scene.call(this,{key: 'gameStats'})
 
     }
 
     preload ()
     {
-		this.load.image('ten', ten)
     }
       
     create ()
     {	
-		
+//these allow usage of attributes and functions from other scenes		
 		var gameStats = this.scene.get('gameStats')
 		var resources = this.scene.get('resources')
 		
-//function that takes the starting parameters from the resources scene and feeds them to the gameStats scene
-		this.startParams = function(params){
-			gameStats.lives = params['lives']
-			gameStats.money = params['money']
-			gameStats.score = params['score']
-			gameStats.speedSetting = params['speedSetting']
-			gameStats.isPlaying = params['isPlaying']
-		}
+//take the starting parameters from the resources scene and feed them to the gameStats scene as different attributes
+//the code is a bit verbose, this shortens it for later use
+		var params = resources.mapData['startParams']
 		
-		gameStats.startParams(resources.mapData['startParams'])
-		
+		this.lives = params['lives']
+		this.money = params['money']
+		this.score = params['score']
+		this.speedSetting = params['speedSetting']
+		this.isPlaying = params['isPlaying']
     }
 	
 	update()
 	{	
-// this allows getting public attributes from the 'map' scene
+//these allow usage of attributes and functions from other scenes
 		var map = this.scene.get('map')
 		var gameStats = this.scene.get('gameStats')
 		var gameRecords = this.scene.get('gameRecords')
 		
-// in the case that the value of lives reaches 0, the scene will change to the 'game over' scene
-		function gameOver(lives, target){
-			if(lives <= 0){
+//in the case that the value of lives reaches 0, the game will change to the game over state
+			if(this.lives <= 0){
 				console.log('game over')
+//record the current score to be displayed later
 				gameRecords.score = gameStats.score
+//call the function to update the high score if needed
 				gameRecords.updateTopScore()
-				
+//record the number of waves survived to be displayed later
 				gameRecords.wavesSurvived = map.origin.waveCounter
 // this changes the scene
 				map.timer = false
-				target.scene.start('overState')
+				this.scene.start('overState')
 			}
-		}
-//this calls the function
-		gameOver(this.lives, this)
-	
-		
-		gameStats.playSpeed = gameStats.speedSetting * gameStats.isPlaying
+//change the speed at which different elements of the game run by multiplying the speed setting by whether the game is meant to be running
+//as the whether the game is meant to be playing is set as a boolean, this means that the multiplication is by either 0 or 1
+		this.playSpeed = this.speedSetting * this.isPlaying
 	}
 }
